Remove tracing_utils_py*
